+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,6 +1,5 @@
-package kitchenpos.product.service;
+package kitchenpos.product.application;
 
-import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.product.dto.ProductRequest;
