+++ src/test/java/kitchenpos/table/domain/OrderTablesTest.java
@@ -1,5 +1,8 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTables;
+import kitchenpos.table.domain.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
